+++ packages/dialtone-vue3/directives/scrollbar_directive/scrollbar.test.js
@@ -9,6 +9,12 @@ const WrapperComponent = {
   `,
 };
 
+const mocks = vi.hoisted(() => {
+  return {
+    destroy: vi.fn(),
+  };
+});
+
 describe('DtScrollbarDirective Tests', () => {
   let wrapper;
   let viewportElement;
@@ -29,13 +35,14 @@ describe('DtScrollbarDirective Tests', () => {
 
   beforeEach(() => {
     OverlayScrollbars.mockClear();
+    mocks.destroy.mockClear();
   });
 
   beforeAll(() => {
     // Mock the overlayscrollbars plugin
     vi.mock('overlayscrollbars', () => {
       const mockPlugin = vi.fn(); // Mock the plugin method
-      const OverlayScrollbarsMock = vi.fn().mockImplementation(() => ({}));
+      const OverlayScrollbarsMock = vi.fn().mockImplementation(() => ({ destroy: mocks.destroy }));
       OverlayScrollbarsMock.plugin = mockPlugin;
       return {
         OverlayScrollbars: OverlayScrollbarsMock,
@@ -66,6 +73,11 @@ describe('DtScrollbarDirective Tests', () => {
         expect(OverlayScrollbars).toHaveBeenCalledTimes(1);
         expect(wrapper.element.getAttribute('data-overlayscrollbars-initialize')).toBe('true');
         expect(wrapper.element.classList.contains('d-scrollbar')).toBe(true);
+      });
+
+      it('should clean up directive', () => {
+        wrapper.unmount();
+        expect(mocks.destroy).toHaveBeenCalledTimes(1);
       });
     });
   });